requirements.txt
# AIMascot Kit - VRM AITuber System
# Python 3.8+ required

# Core Dependencies
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0

# Image & Screenshot
Pillow>=10.0.0
pyobjc>=9.0; sys_platform == "darwin"

# Audio & Speech
SpeechRecognition>=3.10.0
sounddevice>=0.4.6
soundfile>=0.12.1
simpleaudio>=1.0.4
numpy>=1.24.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
pywebview>=4.0.0   # ← 追加 (webviewモジュール)

# Additional APIs
openai>=1.0.0
wikipedia>=1.4.0
